Renamed view module to visualization
